Rename Participant class files
