+++ src/precice/config/ParticipantConfiguration.hpp
@@ -11,7 +11,7 @@
 #include "mesh/SharedPointer.hpp"
 #include "partition/ReceivedPartition.hpp"
 #include "precice/config/SolverInterfaceConfiguration.hpp"
-#include "precice/impl/Participant.hpp"
+#include "precice/impl/ParticipantState.hpp"
 #include "precice/impl/SharedPointer.hpp"
 #include "time/Time.hpp"
 #include "utils/networking.hpp"
