+++ src/sources.cmake
@@ -251,8 +251,8 @@ target_sources(preciceCore
     src/precice/impl/DataContext.hpp
     src/precice/impl/MappingContext.hpp
     src/precice/impl/MeshContext.hpp
-    src/precice/impl/Participant.cpp
-    src/precice/impl/Participant.hpp
+    src/precice/impl/ParticipantState.cpp
+    src/precice/impl/ParticipantState.hpp
     src/precice/impl/ReadDataContext.cpp
     src/precice/impl/ReadDataContext.hpp
     src/precice/impl/SharedPointer.hpp
