+++ tests/serial/mapping-nearest-neighbor-gradient/GradientTestBidirectionalReadScalar.cpp
@@ -22,7 +22,7 @@
 #include "mesh/Vertex.hpp"
 #include "precice/SolverInterface.hpp"
 #include "precice/impl/MeshContext.hpp"
-#include "precice/impl/Participant.hpp"
+#include "precice/impl/ParticipantState.hpp"
 #include "precice/impl/SharedPointer.hpp"
 #include "precice/impl/SolverInterfaceImpl.hpp"
 #include "precice/types.hpp"
